+++ services/autoscaling/tests/unit/test_dynamic_scaling_core.py
@@ -15,8 +15,7 @@ from pydantic import ByteSize, parse_obj_as
 from pytest_mock.plugin import MockerFixture
 from simcore_service_autoscaling.core.settings import ApplicationSettings
 from simcore_service_autoscaling.dynamic_scaling_core import check_dynamic_resources
-from simcore_service_autoscaling.modules.ec2 import get_ec2_client
-from simcore_service_autoscaling.utils_aws import EC2Client
+from types_aiobotocore_ec2.client import EC2Client
 
 
 @pytest.fixture
@@ -26,10 +25,11 @@ def aws_instance_private_dns() -> str:
 
 @pytest.fixture
 def mock_start_aws_instance(
-    mocker: MockerFixture, aws_instance_private_dns: str
+    mocker: MockerFixture,
+    aws_instance_private_dns: str,
 ) -> Iterator[mock.Mock]:
     mocked_start_aws_instance = mocker.patch(
-        "simcore_service_autoscaling.dynamic_scaling_core.utils_aws.start_aws_instance",
+        "simcore_service_autoscaling.modules.ec2.AutoscalingEC2.start_aws_instance",
         autospec=True,
         return_value=aws_instance_private_dns,
     )
@@ -134,7 +134,7 @@ async def test_check_dynamic_resources_with_pending_resources_starts_r5n_4xlarge
 
     await check_dynamic_resources(initialized_app)
     mock_start_aws_instance.assert_called_once_with(
-        get_ec2_client(initialized_app),
+        mock.ANY,
         app_settings.AUTOSCALING_EC2_INSTANCES,
         instance_type="r5n.4xlarge",
         tags=mock.ANY,
