+++ services/autoscaling/tests/unit/test_modules_rabbitmq.py
