+++ services/autoscaling/tests/unit/test_utils_rabbitmq.py
@@ -1,6 +1,7 @@
 # pylint:disable=unused-variable
 # pylint:disable=unused-argument
 # pylint:disable=redefined-outer-name
+# pylint:disable=too-many-arguments
 
 
 from typing import Any, Awaitable, Callable, Mapping
